[PSL-1267] pasteld: optimize historical top mn list in mncache.dat and pasteld
pasteld v2.2.9
List of 10 top masternodes for each block height was saving duplicate masternode information in memory maps, also it was stored in mncache.dat.
Old storage was using map: <block-height> -> <vector of 10 top masternodes with full info>.
Duplicate masternode info was serialized to mncache.dat.
New optimized storage design is using 3 unordered maps:
  1) <block-height> -> <vector of 10 mn cache ids of the top masternodes>
  2) <mn-cache-id> -> <shared_ptr to masternode>
  3) <mn-outpoint> -> <mn-cache-id>
 Map #2 stores unique masternode information, most of the pointers point to the main masternodes storage in MnManager.
 New storage allows to reduce dramatically amount of physical memory required to store historical top mn info.
 Also it allows to reduce mncache.dat size.
 For example:
    - mncache.dat, historical top mn lists: 133644, original size: 780Mb
After optimization:
    - mncache.dat size: 7Mb
	- pasteld memory usage: 3.5Gb -> 2Gb
	- mncache.dat load time on Windows: 40secs -> 1secs
MasterNode Manager serialization version 9.
- Added backward compatibility mode to load previous v8 of the MasterNode Manager,
  converting historical top MNs map to the new storage with 3 maps.
- Added cleanup for duplicate mnbs for the same mn outpoint after mncache load
- Added cleanup for expred mnps after mncache load
- added google tests for mncache.dat, mnpayments.dat
other changes:
 - added new debug log category "wallet", moved all logs related to keypool to "wallet" category

+++ src/mnode/mnode-payments.cpp
@@ -959,8 +959,8 @@ string CMasternodePayments::ToString() const
 {
     ostringstream info;
 
-    info << "Votes: " << (int)mapMasternodePaymentVotes.size() <<
-            ", Blocks: " << (int)mapMasternodeBlockPayees.size();
+    info << "Votes: " << mapMasternodePaymentVotes.size() <<
+            ", Blocks: " << mapMasternodeBlockPayees.size();
 
     return info.str();
 }
